requirements.txt
flask==2.2.2
requests==2.28.1
psycopg2-binary==2.9.5
mysql-connector-python==8.0.31
pandas==1.3.5
pymongo==4.3.3
pytest==7.1.3
flask_cors==3.0.10